// src/generated/patentsview/zod.gen.ts
export const zApiSuccessResponse = z.object({
  error: z.boolean(),
  count: z.int(),
  total_hits: z.int(),
})

export const zGBrfSumTextSuccessResponse = zApiSuccessResponse.and(
  z.object({
    g_brf_sum_texts: z
      .array(
        z.object({
          patent_id: z.string().optional(),
          summary_text: z.string().optional(),
        }),
      )
      .optional(),
  }),
)

export const zGClaimSuccessResponse = zApiSuccessResponse.and(
  z.object({
    g_claims: z
      .array(
        z.object({
          patent_id: z.string().optional(),
          claim_sequence: z.int().optional(),
          claim_number: z.string().optional(),
          claim_text: z.string().optional(),
          exemplary: z.int().optional(),
          claim_dependent: z.string().optional(),
        }),
      )
      .optional(),
  }),
)
// ...
export const zAssigneeSuccessResponse = zApiSuccessResponse.and(
  z.object({
    assignees: z
      .array(
        z.object({
          assignee_id: z.string().optional(),
          assignee_individual_name_first: z.string().optional(),
          assignee_individual_name_last: z.string().optional(),
          assignee_organization: z.string().optional(),
          assignee_type: z.string().optional(),
          assignee_lastknown_city: z.string().optional(),
          assignee_lastknown_state: z.string().optional(),
          assignee_lastknown_country: z.string().optional(),
          assignee_lastknown_latitude: z.string().optional(),
          assignee_lastknown_longitude: z.string().optional(),
          assignee_lastknown_location: z.string().optional(),
          assignee_first_seen_date: z.iso.date().optional(),
          assignee_last_seen_date: z.iso.date().optional(),
          assignee_num_inventors: z.number().optional(),
          assignee_num_patents: z.number().optional(),
          assignee_years_active: z.number().optional(),
          assignee_years: z
            .array(
              z.object({
                year: z.number().optional(),
                num_patents: z.number().optional(),
              }),
            )
            .optional(),
        }),
      )
      .optional(),
  }),
)

export const zAttorneySuccessResponse = zApiSuccessResponse.and(
  z.object({
    attorneys: z
      .array(
        z.object({
          attorney_id: z.string().optional(),
          attorney_name_first: z.string().optional(),
          attorney_name_last: z.string().optional(),
          attorney_organization: z.string().optional(),
          attorney_first_seen_date: z.iso.date().optional(),
          attorney_last_seen_date: z.iso.date().optional(),
          attorney_num_inventors: z.number().optional(),
          attorney_num_patents: z.number().optional(),
          attorney_years_active: z.number().optional(),
        }),
      )
      .optional(),
  }),
)
// ...
export const zCpcSubclassSuccessResponse = zApiSuccessResponse.and(
  z.object({
    cpc_subclasses: z
      .array(
        z.object({
          cpc_subclass_id: z.string().optional(),
          cpc_subclass_title: z.string().optional(),
          cpc_class_id: z.string().optional(),
          cpc_class: z.string().optional(),
          cpc_subclass_first_seen_date: z.iso.date().optional(),
          cpc_subclass_last_seen_date: z.iso.date().optional(),
          cpc_subclass_years_active: z.number().optional(),
          cpc_subclass_num_patents: z.number().optional(),
          cpc_subclass_num_assignees: z.number().optional(),
          cpc_subclass_num_inventors: z.number().optional(),
        }),
      )
      .optional(),
  }),
)
// ...
export const zInventorSuccessResponse = zApiSuccessResponse.and(
  z.object({
    inventors: z
      .array(
        z.object({
          inventor_id: z.string().optional(),
          inventor_name_first: z.string().optional(),
          inventor_name_last: z.string().optional(),
          inventor_gender_code: z.string().optional(),
          inventor_lastknown_city: z.string().optional(),
          inventor_lastknown_state: z.string().optional(),
          inventor_lastknown_country: z.string().optional(),
          inventor_lastknown_latitude: z.string().optional(),
          inventor_lastknown_longitude: z.string().optional(),
          inventor_lastknown_location: z.string().optional(),
          inventor_first_seen_date: z.iso.date().optional(),
          inventor_last_seen_date: z.iso.date().optional(),
          inventor_num_assignees: z.number().optional(),
          inventor_num_patents: z.number().optional(),
          inventor_years_active: z.number().optional(),
          inventor_years: z
            .array(
              z.object({
                year: z.number().optional(),
                num_patents: z.number().optional(),
              }),
            )
            .optional(),
        }),
      )
      .optional(),
  }),
)

export const zIpcClassificationSuccessResponse = zApiSuccessResponse.and(
  z.object({
    ipcr: z
      .array(
        z.object({
          ipc_id: z.string().optional(),
          ipc_section: z.string().optional(),
          ipc_class: z.string().optional(),
          ipc_subclass: z.string().optional(),
        }),
      )
      .optional(),
  }),
)